1. fingersUp-OpenCV/mqtt/requirements.txt
opencv-python==4.6.0.66
cvzone==1.4.1
mediapipe==0.8.1
paho-mqtt==1.6.1